# server/swagger-calls-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Calls & SIP Integration API",
    "description": "Unified call management API for MediLead CMS. Supports telephony vendor webhooks (HMAC-SHA256), SIP event processing, manual call logging, real-time Socket.IO events, and intercom-based agent routing.\n\n## Authentication\n- **POST /api/calls/inbound**: HMAC-SHA256 via `X-Webhook-Signature` header, or JWT Bearer fallback\n- **All other endpoints**: JWT Bearer token via `Authorization` header\n\n## Base URL\n`http://localhost:5000/api`",
    "version": "1.0.0",
    "contact": {
      "name": "MediLead CMS Support",
      "url": "https://medilead.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/calls": {
      "get": {
        "summary": "List call logs with pagination and filters",
        "description": "Returns paginated call logs with optional filters. Non-admin/non-manager users only see calls matching their intercom_number. Supports both `status` and `call_status` query params for backward compatibility.",
        "tags": ["Calls"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "user_id", "in": "query", "schema": { "type": "integer" }, "description": "Filter by assigned user ID" },
          { "name": "direction", "in": "query", "schema": { "type": "string", "enum": ["inbound", "outbound"] }, "description": "Filter by call direction" },
          { "name": "status", "in": "query", "schema": { "type": "string" }, "description": "Filter by call status (alias for call_status)" },
          { "name": "call_status", "in": "query", "schema": { "type": "string", "enum": ["initiated", "ringing", "in-progress", "completed", "failed", "missed"] }, "description": "Filter by call status" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of calls with lead and user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "calls": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CallLog" }
                        },
                        "total": { "type": "integer", "description": "Total number of matching records" },
                        "page": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid JWT token" },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CallsFetchError" } } } }
        }
      },
      "post": {
        "summary": "Log a call manually",
        "description": "Creates a manual call log entry. Auto-links lead by caller_number if lead_id not provided. Updates lead.last_call_date for linked calls. Sends missed-call notification for missed status.",
        "tags": ["Calls"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["caller_number"],
                "properties": {
                  "caller_number": { "type": "string", "pattern": "^[0-9]{10,15}$", "description": "Caller phone number (10-15 digits)", "example": "9876543210" },
                  "callee_number": { "type": "string", "pattern": "^[0-9]{10,15}$", "description": "Callee phone number (10-15 digits)", "example": "9988776655" },
                  "direction": { "type": "string", "enum": ["inbound", "outbound"], "default": "inbound", "description": "Call direction" },
                  "status": { "type": "string", "default": "initiated", "description": "Call status" },
                  "duration": { "type": "integer", "minimum": 0, "default": 0, "description": "Call duration in seconds" },
                  "lead_id": { "type": "integer", "description": "Lead ID to link (skips auto-lookup if provided)" },
                  "notes": { "type": "string", "maxLength": 2000, "description": "Call notes" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Call logged successfully", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" }, "data": { "type": "object", "properties": { "call": { "$ref": "#/components/schemas/ManualCallLog" } } } } } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CallLogError" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth/login"
      },
      "HmacSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Signature",
        "description": "HMAC-SHA256 hex digest of the raw request body, signed with TELEPHONY_WEBHOOK_SECRET"
      }
    },
    "schemas": {
      "CallLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "caller_number": { "type": "string" },
          "callee_number": { "type": "string", "nullable": true },
          "direction": { "type": "string" },
          "call_status": { "type": "string" },
          "status": { "type": "string", "description": "Backward-compatible alias for call_status" },
          "duration": { "type": "integer" },
          "duration_seconds": { "type": "integer" },
          "recording_url": { "type": "string", "nullable": true },
          "intercom_number": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" },
          "received_at": { "type": "string", "format": "date-time" },
          "created_at": { "type": "string", "format": "date-time" },
          "vendor_call_id": { "type": "string" },
          "code": { "type": "string" },
          "lead_id": { "type": "integer", "nullable": true },
          "lead_name": { "type": "string", "nullable": true },
          "lead_phone": { "type": "string", "nullable": true },
          "user_id": { "type": "integer", "nullable": true },
          "user_name": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "ManualCallLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "caller_number": { "type": "string" },
          "direction": { "type": "string" },
          "call_status": { "type": "string" },
          "status": { "type": "string" },
          "duration": { "type": "integer" },
          "lead_id": { "type": "integer", "nullable": true },
          "user_id": { "type": "integer" },
          "notes": { "type": "string", "nullable": true },
          "code": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CallsFetchError": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "error" },
          "message": { "type": "string" },
          "code": { "type": "string", "example": "CALLS_FETCH_ERROR" }
        }
      },
...
      "CallLogError": { "type": "object", "properties": { "status": { "type": "string" }, "message": { "type": "string" }, "code": { "type": "string", "example": "CALL_LOG_ERROR" } } },
...
    }
  }
}